requirements.txt
# Core web framework
flask>=2.3.0
requests>=2.28.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-flask>=1.0.0

# Cassandra database
cassandra-driver>=3.25.0

# Text processing
nltk>=3.8.0

# Utilities
python-json-logger>=2.0.0